+++ src/Image/Sanitizer/Detector/ForeignObjectDetector.php
@@ -1,0 +1,24 @@
+<?php
+
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+declare(strict_types=1);
+
+namespace OxidEsales\MediaLibrary\Image\Sanitizer\Detector;
+
+use DOMDocument;
+use DOMXPath;
+
+final class ForeignObjectDetector implements SvgViolationDetectorInterface
+{
+    public function detect(DOMDocument $document): bool
+    {
+        $xpath = new DOMXPath($document);
+        $nodes = $xpath->query('//*[local-name()="foreignObject"]');
+
+        return $nodes !== false && $nodes->length > 0;
+    }
+}